Added in tests for payment types
