+++ bangazon_control.py
@@ -46,7 +46,7 @@ class BangazonControl(object):
     def choose_active_customer(self):
         return 1
 
-    def create_payment_type(self, name, account_number):
+    def create_payment_type(self, customer_id, name, account_number):
         """
         This method creates a dictionary that contains information about the payment types and returns that dictionary.
 
@@ -60,5 +60,16 @@ class BangazonControl(object):
         Author:
             wocaldwell
         """
-        new_payment_type = {'name': name, 'account_number': account_number}
+        new_payment_type = {'customer_id': customer_id, 'name': name, 'account_number': account_number}
         return new_payment_type
+
+    def save_all_customers(self):
+        pass
+
+
+
+
+
+
+
+
