+++ src/product_order.py
@@ -1,1 +1,14 @@
-#.
+"""
+Terminal Interface configuration for the ProductOrder table interaction in the database
+"""
+class ProductOrder(object):
+    """
+    This class is to handle interactions with the ProductOrder table in database through sqlite
+
+    Methods:        
+
+    Author:
+        Adam Myers
+    """
+
+
